requirements.txt
fastapi
uvicorn
prisma
python-dotenv
pyjwt
httpx
pytest